requirements-dev.txt
pytest>=8.0
PyYAML>=6.0
tomli>=2.0; python_version < "3.11"
